Move the rest of harness implementation into "harness" package
This is make it live in other than "core" or "org.renaissance",
both of which are loaded by a different class loader. No further
sweeping changes are planned :-)
