+++ renaissance-harness/src/main/scala/org/renaissance/harness/ConfigParser.scala
@@ -1,4 +1,4 @@
-package org.renaissance
+package org.renaissance.harness
 
 import scopt.OptionParser
